Transport problem
There is a certain uniform cargo, which needs to be transported from n warehouses to m plants. For each warehouse i, it is known how much cargo ai is in it, and for each plant its demand bj in cargo is known. The cost of transportation is proportional to the distance from the warehouse to the factory (all distances cij from the i-th warehouse to the j-th plant are known). It is required to make the cheapest plan of transportation.